+++ src/Ch03/Nullable.hs
@@ -12,14 +12,18 @@
 
 module Ch03.Nullable where
 
-import Prelude hiding (Just)
+import Prelude hiding (Maybe, Just)
 
 -- Parameterized Types, p. 57
 
 data Maybe a = Just a
              | Nothing
 
+someBool :: Maybe Bool
 someBool = Just True
+
+someString :: Maybe String
 someString = Just "something"
 
+wrapped :: Maybe (Maybe String)
 wrapped = Just (Just "wrapped")